4997 move functions to more appropriate location
