+++ website/src/components/SequenceDetailsPage/SequencesDisplay/getSequenceNames.tsx
@@ -1,5 +1,11 @@
 import { type ReferenceGenomesSequenceNames, SINGLE_REFERENCE } from '../../../types/referencesGenomes.ts';
-import type { SequenceName } from '../../../utils/sequenceTypeHelpers.ts';
+import {
+    getMultiPathogenNucleotideSequenceNames,
+    getMultiPathogenSequenceName,
+    getSinglePathogenSequenceName,
+    isMultiSegmented,
+    type SequenceName,
+} from '../../../utils/sequenceTypeHelpers.ts';
 
 export function getSequenceNames(
     referenceGenomeSequenceNames: ReferenceGenomesSequenceNames,
@@ -25,28 +31,4 @@ export function getSequenceNames(
         genes: genes.map((name) => getMultiPathogenSequenceName(name, suborganism)),
         isMultiSegmented: true, // LAPIS treats the suborganisms as multiple nucleotide segments -> always true
     };
-}
-
-export function getMultiPathogenNucleotideSequenceNames(nucleotideSequences: string[], suborganism: string) {
-    return nucleotideSequences.length === 1
-        ? [{ lapisName: suborganism, label: 'main' }]
-        : nucleotideSequences.map((name) => getMultiPathogenSequenceName(name, suborganism));
-}
-
-export function getSinglePathogenSequenceName(name: string): SequenceName {
-    return {
-        lapisName: name,
-        label: name,
-    };
-}
-
-export function getMultiPathogenSequenceName(name: string, suborganism: string): SequenceName {
-    return {
-        lapisName: `${suborganism}-${name}`,
-        label: name,
-    };
-}
-
-export function isMultiSegmented(nucleotideSegmentNames: unknown[]) {
-    return nucleotideSegmentNames.length > 1;
 }
