+++ website/src/utils/sequenceTypeHelpers.ts
@@ -8,6 +8,30 @@ export type SequenceName = {
     label: string;
 };
 
+export function getMultiPathogenNucleotideSequenceNames(nucleotideSequences: string[], suborganism: string) {
+    return nucleotideSequences.length === 1
+        ? [{ lapisName: suborganism, label: 'main' }]
+        : nucleotideSequences.map((name) => getMultiPathogenSequenceName(name, suborganism));
+}
+
+export function getSinglePathogenSequenceName(name: string): SequenceName {
+    return {
+        lapisName: name,
+        label: name,
+    };
+}
+
+export function getMultiPathogenSequenceName(name: string, suborganism: string): SequenceName {
+    return {
+        lapisName: `${suborganism}-${name}`,
+        label: name,
+    };
+}
+
+export function isMultiSegmented(nucleotideSegmentNames: unknown[]) {
+    return nucleotideSegmentNames.length > 1;
+}
+
 export const unalignedSequenceSegment = (segmentName: SequenceName): SequenceType => ({
     type: 'nucleotide',
     aligned: false,
